+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog][keepachangelog], and this project adhe
 
 - Allow configuring the behavior for failures in method argument resolution (mostly deserialization) for
   `@JetStreamListener`-annotated methods.
+- Allow `@NatsHeader`-annotated parameter without `value` attribute (fallback to method argument name).
 
 ## [0.3.0] - 2026-06-13